pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "versed"
version = "0.0.7"
description = ''
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = []
authors = [
  { name = "lachiewalker", email = "lachiewalker1@hotmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "textual",
    "pymilvus",
    "milvus-lite",
    "google-auth-oauthlib",
    "google-api-python-client",
    "platformdirs",
    "keyring",
    "cryptography",
    "openai",
    "python-docx",
    "langchain-text-splitters"
]

[project.urls]
Documentation = "https://github.com/lachiewalker/versed#readme"
Issues = "https://github.com/lachiewalker/versed/issues"
Source = "https://github.com/lachiewalker/versed"

[project.scripts]
versed = "versed.__main__:cli"

[tool.hatch.version]
path = "pyproject.toml"

[tool.hatch.build]
artifacts = [
  "credentials.json",
]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/versed tests}"

[tool.coverage.run]
source_pkgs = ["versed", "tests"]
branch = true
parallel = true
omit = [
  "src/versed/__about__.py",
]

[tool.coverage.paths]
versed = ["src/versed", "*/versed/src/versed"]
tests = ["tests", "*/versed/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
